Change to double

+++ src/main/java/com/aetherwars/controller/CardController.java
@@ -34,8 +34,8 @@ public class CardController {
     * Display character card information for character card*/
     public void setCard(Character character){
         Image newImg = new Image("/com/aetherwars/model/card/character/image/" + character.getImagepath().toString());
-        String atk = Integer.toString(character.getBaseAttack());
-        String hp = Integer.toString(character.getBaseHealth());
+        String atk = Double.toString(character.getBaseAttack());
+        String hp = Double.toString(character.getBaseHealth());
         String desc = "ATK " + atk + "/HP " + hp;
 
         cardImage.setImage(newImg);
@@ -61,8 +61,8 @@ public class CardController {
      * Display character card information for potion spell card*/
     public void setCard(Potion potion){
         Image newImg = new Image("/com/aetherwars/model/card/spell/potion/image/" + potion.getImagepath().toString());
-        String atk = Integer.toString(potion.getTempAttack());
-        String hp = Integer.toString(potion.getTempHealth());
+        String atk = Double.toString(potion.getTempAttack());
+        String hp = Double.toString(potion.getTempHealth());
         String desc = "ATK + " + atk + "/HP +" + hp;
 
         cardImage.setImage(newImg);
